Saving an entire goodness produced by claude
The prompt was:  I want to explore  feature of git -- git namespaces and their applicability of containing entire hierarchies of git submodules in the same
repository. The goal would be to e.g. try to define namespaces per each nested submodule and see if we can easily clone from the same github
repository but for a given namespace and see if could be specified somehow within .gitmodules that the url with specific namespace. Also that
when we clone from main namespace we do not get all the objects for submodule. Explore this feature of git namespaces locally in this folder,
and then can also experiment with https://github.com/yarikoptic/test-git-namespaces repository